# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Система противодействия отмыванию денег и мошенничеству",
        "title": "Bank AML System API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/transactions": {
            "get": {
                "description": "Возвращает список всех транзакций с пагинацией",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Получить список транзакций",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Лимит результатов (максимум 500)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список транзакций",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Принимает транзакцию и отправляет её на анализ рисков через REST API. Транзакция сохраняется в БД и отправляется в Kafka для асинхронной обработки fraud-сервисом.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Отправить транзакцию на анализ (REST)",
                "parameters": [
                    {
                        "description": "Данные транзакции",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bank-aml-system_internal_models.ProcessingRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Транзакция принята на обработку",
                        "schema": {
                            "$ref": "#/definitions/bank-aml-system_internal_models.ProcessingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет все транзакции из базы данных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Очистить все транзакции",
                "responses": {
                    "200": {
                        "description": "Транзакции очищены",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/transactions/generate": {
            "get": {
                "description": "Генерирует случайную транзакцию для тестирования",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Сгенерировать случайную транзакцию",
                "responses": {
                    "200": {
                        "description": "Сгенерированная транзакция",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/transactions/grpc": {
            "post": {
                "description": "Принимает транзакцию и отправляет её на анализ через gRPC-сервис. Транзакция автоматически сохраняется в БД, отправляется в Kafka, обрабатывается fraud-сервисом и возвращается с результатами анализа рисков (risk_score, risk_level, flags).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Отправить транзакцию через gRPC",
                "parameters": [
                    {
                        "description": "Данные транзакции",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bank-aml-system_internal_models.ProcessingRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Транзакция успешно обработана через gRPC",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request - неверный формат данных",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - ошибка обработки",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - gRPC клиент недоступен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/transactions/{processing_id}": {
            "get": {
                "description": "Возвращает детальную информацию о транзакции и её анализе рисков",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Получить статус транзакции",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID обработки транзакции",
                        "name": "processing_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Статус транзакции",
                        "schema": {
                            "$ref": "#/definitions/bank-aml-system_internal_models.TransactionStatusResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "bank-aml-system_internal_models.ProcessingRequest": {
            "type": "object",
            "required": [
                "account_number",
                "amount",
                "currency",
                "transaction_id",
                "transaction_type"
            ],
            "properties": {
                "account_number": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                },
                "branch_id": {
                    "type": "string"
                },
                "channel": {
                    "type": "string"
                },
                "counterparty_account": {
                    "type": "string"
                },
                "counterparty_bank": {
                    "type": "string"
                },
                "counterparty_country": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                },
                "transaction_type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "bank-aml-system_internal_models.ProcessingResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "processing_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "bank-aml-system_internal_models.TransactionStatusResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "analysis_timestamp": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "flags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "processing_id": {
                    "type": "string"
                },
                "risk_level": {
                    "type": "string"
                },
                "risk_score": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                }
            }
        }
    }
}